% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{fix_analyte_names}
\alias{fix_analyte_names}
\title{Fix SomaScan analyte IDs}
\usage{
fix_analyte_names(x)
}
\arguments{
\item{x}{Character vector of analyte IDs.}
}
\value{
Character vector with \verb{_} converted to \code{.} for analyte-like tokens.
}
\description{
Converts analyte IDs of the form \code{X10000_28} to \code{X10000.28}.
Leaves other strings unchanged.
}
