% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_panels.R
\name{load_modules_fixed}
\alias{load_modules_fixed}
\title{Load a fixed modules artifact from disk}
\usage{
load_modules_fixed(path)
}
\arguments{
\item{path}{Path to the RDS created by \code{save_modules_fixed()}.}
}
\value{
The modules object.
}
\description{
Load a fixed modules artifact from disk
}
